refactor: fix build & build:docs
