+++ .eslintignore
@@ -1,3 +1,4 @@
 /src/typechain/*
+**/*.test.ts
 node_modules/*
 /dist
