+++ docs/modules.md
@@ -5,6 +5,7 @@
 -   [IPFS](modules/IPFS.md)
 -   [LSP12](modules/LSP12.md)
 -   [LSP2](modules/LSP2.md)
+-   [LSP23](modules/LSP23.md)
 -   [LSP3](modules/LSP3.md)
 -   [LSP4](modules/LSP4.md)
 -   [LSP5](modules/LSP5.md)
