+++ docs/modules/LSP23.md
@@ -1,0 +1,34 @@
+# Module: LSP23
+
+## LSP23
+
+### deployUniversalProfile
+
+▸ **deployUniversalProfile**(`signer`, `mainController`, `additionalDataKeys`, `additionalDataValues`, `additionalControllers`, `salt?`): `Promise`\<\{ `keyManagerAddress`: `string` ; `universalProfileAddress`: `string` }\>
+
+Deploy Universal Profile with Key Manager using LSP23 Linked Contracts Factory.
+
+#### Parameters
+
+| Name                    | Type                 | Description                                                                       |
+| :---------------------- | :------------------- | :-------------------------------------------------------------------------------- |
+| `signer`                | `Signer` \| `Wallet` | The signer used to deploy the contracts, needs to have native token for gas fees. |
+| `mainController`        | `BytesLike`          | The main controller for the Unviersal Profile.                                    |
+| `additionalDataKeys`    | `BytesLike`[]        | Data keys that you want to set on deployment.                                     |
+| `additionalDataValues`  | `BytesLike`[]        | Data values that you want to set on deployment.                                   |
+| `additionalControllers` | `LSP6Controller`[]   | Additional controllers for the deployed Universal Profile.                        |
+| `salt?`                 | `BytesLike`          | -                                                                                 |
+
+#### Returns
+
+`Promise`\<\{ `keyManagerAddress`: `string` ; `universalProfileAddress`: `string` }\>
+
+The addresses of the Universal Profile & Key Manager.
+
+**`Since`**
+
+v0.0.2
+
+#### Defined in
+
+[LSP23/deployUniversalProfile/deployUniversalProfile.ts:34](https://github.com/lukso-network/lsp-utils/blob/main/src/LSP23/deployUniversalProfile/deployUniversalProfile.ts#L34)
