+++ package.json
@@ -26,10 +26,9 @@
         "lint": "eslint . --ext .ts",
         "prettier": "prettier -w .",
         "clean": "rm -rf dist && rm -rf src/typechain",
-        "build:docs": "typedoc --options typedoc.json",
-        "build:types:all": "typechain --target=ethers-v6 --out-dir src/typechain ./node_modules/@lukso/lsp-smart-contracts/artifacts/*.json ./node_modules/@erc725/smart-contracts/artifacts/*.json ./node_modules/@openzeppelin/contracts/build/contracts/*.json",
-        "build:types:specific": "bash scripts/generateSpecificTypechain.bash",
-        "build": "npm run clean && npm run lint && npm run build:types:specific && vite build",
+        "build:docs": "typedoc --options typedoc.json && prettier -w ./docs",
+        "build:types": "bash scripts/generateTypechain.bash",
+        "build": "npm run clean && npm run lint && npm run build:types && vite build",
         "test:IPFS": "hardhat test src/IPFS/**/*.test.ts",
         "test:LSP2": "hardhat test src/LSP2/**/*.test.ts",
         "test:LSP3": "hardhat test src/LSP3/**/*.test.ts",
