+++ scripts/compareSrcAndTests.bash
@@ -12,6 +12,12 @@ do
         continue
     fi;
 
+    # ignore src/***/**/index.ts
+    if [[ $utilPath == src/typechain/* ]];
+    then
+        continue
+    fi;
+
     # generate the `testPath` from `utilPath`
     ## remove the longest string from right to left
     ## which starts with "."
