+++ src/LSP23/deployUniversalProfile/deployUniversalProfile.test.ts
@@ -1,3 +1,4 @@
+// TODO Find a way to test contract deployemnt without hardhat environment
 describe('deployUniversalProfile', () => {
     it.skip('should deploy and set data keys and data values correctly', async () => {});
 });
