+++ src/LSP4/getDigitalAssetCreators/getDigitalAssetCreators.ts
@@ -1,8 +1,8 @@
 import { BytesLike, Provider, isAddress, isHexString, toNumber } from 'ethers';
 import { ERC725YDataKeys } from '@lukso/lsp-smart-contracts';
 
-// types
-import { ERC725Y } from '../../typechain';
+// typechain
+import { ERC725Y } from '../../typechain/erc725';
 
 // constants
 import {
@@ -25,7 +25,7 @@ import {
  *
  * @throws
  * - When `digitalAssetAddress` is not a valid address.
- * - When the contract deployed at `digitalAssetAddress` address does not support the `ERC725Y` interface id.
+ * - When the contract deployed at `digitalAssetAddress` does not support the `ERC725Y` interface id.
  * - When the length for `LSP4Creators[]` is not a valid LSP2 array length value.
  *
  * @see https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-4-DigitalAsset-Metadata.md
