--- tests/LSP23/deployUniversalProfile.test.ts
@@ -1,3 +1,0 @@
-describe('deployUniversalProfile', () => {
-    it('WIP', () => {});
-});
